Remove profile template and seeding infrastructure
Strips out the entire profile configuration system including templates,
sections, styles, and data seeding mechanisms.

Removes unused services (ProfileService, TemplateService,
ReadmeGeneratorService) and their associated interfaces, models, and
DTOs to simplify the codebase.

Also fixes GitHub ID claim parsing to use long instead of int and
removes the Languages property from GitHubRepository model.

+++ src/Profily.Infrastructure/Services/AuthService.cs
@@ -22,7 +22,7 @@ public sealed class AuthService : IAuthService
         var githubIdClaim = principal.FindFirst(ClaimTypes.NameIdentifier)?.Value
             ?? throw new InvalidOperationException("GitHub ID claim not found.");
         
-        var githubId = int.Parse(githubIdClaim);
+        var githubId = long.Parse(githubIdClaim);
         var username = principal.FindFirst(ClaimTypes.Name)?.Value
             ?? throw new InvalidOperationException("GitHub username claim not found.");
         var email = principal.FindFirst(ClaimTypes.Email)?.Value;
